Fix implicit int to float cast

+++ local_planner/include/local_planner/planner_functions.h
@@ -26,7 +26,7 @@ namespace avoidance {
 * @param[in]  FOV, struct defining current field of view
 * @param[in]  position, current vehicle position
 * @param[in]  min_realsense_dist, minimum sensor range [m]
-* @param[in]  max_age, maximum age (compute cycles) to keep data
+* @param[in]  max_age, maximum age in seconds to keep data
 * @param[in]  elapsed, time elapsed since last processing [s]
 * @param[in]  min_num_points_per_cell, number of points from which on they will
 *             be kept, less points are discarded as noise (careful: 0 is not
@@ -35,7 +35,7 @@ namespace avoidance {
 void processPointcloud(pcl::PointCloud<pcl::PointXYZI>& final_cloud,
                        const std::vector<pcl::PointCloud<pcl::PointXYZ>>& complete_cloud, const Box& histogram_box,
                        const std::vector<FOV>& fov, float yaw_fcu_frame_deg, float pitch_fcu_frame_deg,
-                       const Eigen::Vector3f& position, float min_realsense_dist, int max_age, float elapsed_s,
+                       const Eigen::Vector3f& position, float min_realsense_dist, float max_age, float elapsed_s,
                        int min_num_points_per_cell);
 
 /**
